refactor(arenamaster): split login into files
